refactor: . Move adjustRelativeLinksInDescription() to TaskLineRenderer.ts
